python/year-service/pyproject.toml
[tool.poetry]
name = "year-service"
version = "0.1.0"
description = ""
authors = ["Honeycomb"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
python-dotenv = "^1.0.1"
opentelemetry-instrumentation-django = "^0.44b0"
opentelemetry-sdk = "^1.23.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.23.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.2"
autopep8 = "^2.0.4"
